refactor: Move splitMessageString function to helpers
